requirements.txt
authlib>=1.6.1
cryptography>=45.0.6
fastapi>=0.116.1
httpx>=0.28.1
itsdangerous>=2.2.0
mcp[cli]>=1.13.0
pathlib>=1.0.1
pyjwt[crypto]>=2.10.1
python-dotenv>=1.1.1
requests>=2.32.5
sse-starlette>=3.0.2
uvicorn>=0.35.0